+++ teled/main.go
@@ -25,7 +25,6 @@ func main() {
 			"fwd":           true,
 			"edit":          true,
 			"like":          true,
-			"dislike":       true,
 		},
 		Start: func(ctx context.Context, rc *chanlib.RouterClient) (http.Handler, func(), error) {
 			b, err := newBot(cfg)